restore fail2ban

+++ add-fail2ban.sh
@@ -1,18 +1,10 @@
 #!/bin/bash
-#######################################################################################################################
+##########################################################################
 # Add fail2ban restrictions to Guacamole
 # For Ubuntu / Debian / Raspian
 # David Harrop
 # April 2023
-#######################################################################################################################
-
-clear
-
-# Check if user is root or sudo
-if ! [ $( id -u ) = 0 ]; then
-	echo -e "${LGREEN}Please run this script as sudo or root${NC}" 1>&2
-	exit 1
-fi
+##########################################################################
 
 # Prepare text output colours
 GREY='\033[0;37m'
@@ -32,12 +24,19 @@ MAGENTA='\033[0;35m'
 LMAGENTA='\033[0;95m'
 NC='\033[0m' #No Colour
 
+clear
+
+# Check if user is root or sudo
+if ! [ $( id -u ) = 0 ]; then
+	echo -e "${LGREEN}Please run this script as sudo or root${NC}" 1>&2
+	exit 1
+fi
+
 # Initialise variables
 FAIL2BAN_BASE=""
 FAIL2BAN_GUAC=""
 FAIL2BAN_NGINX=""
 FAIL2BAN_SSH=""
-TOMCAT="tomcat9"
 
 #Clean up from any previous runs
 rm -f /tmp/fail2ban.conf
@@ -45,10 +44,7 @@ rm -f /tmp/ip_list.txt
 rm -f /tmp/netaddr.txt
 rm -f /tmp/fail2ban.update
 
-
-#######################################################################################################################
-# Begin install menu prompts ##########################################################################################
-#######################################################################################################################
+####################### Start installation prompts #######################
 
 # Prompt to install fail2ban base app, default of yes
 if [[ -z ${FAIL2BAN_BASE} ]]; then
@@ -94,11 +90,9 @@ if [[ -z ${FAIL2BAN_SSH} ]] && [[ "${FAIL2BAN_BASE}" = true ]]; then
 	FAIL2BAN_SSH=false
 	fi
 fi
+######################## End installation prompts ########################
 
-
-#######################################################################################################################
-# Start fail2ban global setup #########################################################################################
-#######################################################################################################################
+############ Start Fail2ban base application install option ##############
 
 # Install base fail2ban base application (no policy defined yet)
 if [ "${FAIL2BAN_BASE}" = true ]; then
@@ -216,12 +210,21 @@ echo -e "${LGREEN}Fail2ban setup cancelled.${GREY}"
 
 fi
 
+############## End Fail2ban base application install option ##############
 
-#######################################################################################################################
-# Start optional setup actions   ######################################################################################
-#######################################################################################################################
 
-### Fail2ban Guacamole security policy option ###
+########### Start Fail2ban Guacamole security policy option ##############
+
+if [ "${FAIL2BAN_GUAC}" = true ]; then
+# Find the correct tomcat package (with a little future proofing) so we can configure the fail2ban log path
+if [[ $( apt-cache show tomcat10 2> /dev/null | egrep "Version: 10" | wc -l ) -gt 0 ]]; then
+	TOMCAT="tomcat10"
+	elif [[ $( apt-cache show tomcat9 2> /dev/null | egrep "Version: 9" | wc -l ) -gt 0 ]]; then
+	TOMCAT="tomcat9"
+else
+	echo -e "${RED}Failed. Can't find Tomcat package${GREY}" 1>&2
+	exit 1
+fi
 
 # Create the Guacamole jail.local policy template
 cat > /tmp/fail2ban.conf <<EOF
@@ -262,13 +265,13 @@ echo -e "${LGREEN}Guacamole security policy applied${GREY}\n-${SED_NETADDR}are w
 echo
 fi
 
-### Fail2ban NGINX security policy option ###
+############## Start Fail2ban NGINX security policy option ###############
 if [ "${FAIL2BAN_NGINX}" = true ]; then
 echo -e "${LGREEN}Nginx Fail2ban policy not implemented yet.${GREY}"
 echo
 fi
 
-### Start Fail2ban SSH security policy option ###
+############### Start Fail2ban SSH security policy option ################
 if [ "${FAIL2BAN_SSH}" = true ]; then
 echo -e "${LGREEN}SSH Fail2ban policy not implemented yet..${GREY}"
 echo
